Restructure VM
